+++ client/src/lib/form-validators/name-length-validator.js
@@ -1,0 +1,5 @@
+import { createCharacterLengthRange } from '@dhis2/ui'
+
+const twoToHundredValidator = createCharacterLengthRange(2, 100)
+
+export const nameLengthValidator = value => twoToHundredValidator(value)