+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [2.28.4](https://github.com/dhis2/app-hub/compare/v2.28.3...v2.28.4) (2021-12-03)
+
+
+### Bug Fixes
+
+* sort versions by created-date ([#577](https://github.com/dhis2/app-hub/issues/577)) ([0addf38](https://github.com/dhis2/app-hub/commit/0addf38dd7c20e6072fbb75258a6c3c0e0cfe534))
+* validate app-name length ([#572](https://github.com/dhis2/app-hub/issues/572)) ([c6857eb](https://github.com/dhis2/app-hub/commit/c6857ebafd51bf198764aff6df2166a083e83b45))
+
 ## [2.28.3](https://github.com/dhis2/app-hub/compare/v2.28.2...v2.28.3) (2021-12-01)
 
 
@@ -87,6 +95,19 @@
 ### Features
 
 * add coreApp property ([#522](https://github.com/dhis2/app-hub/issues/522)) ([e22dafd](https://github.com/dhis2/app-hub/commit/e22dafdbb561bf248204079607727c933a6f58cd))
+
+## [2.23.9](https://github.com/dhis2/app-hub/compare/v2.23.8...v2.23.9) (2021-10-29)
+
+
+### Bug Fixes
+
+* sort versions by created-date ([#577](https://github.com/dhis2/app-hub/issues/577)) ([0addf38](https://github.com/dhis2/app-hub/commit/0addf38dd7c20e6072fbb75258a6c3c0e0cfe534))
+
+## [2.23.8](https://github.com/dhis2/app-hub/compare/v2.23.7...v2.23.8) (2021-10-15)
+
+### Bug Fixes
+
+* validate app-name length ([#572](https://github.com/dhis2/app-hub/issues/572)) ([c6857eb](https://github.com/dhis2/app-hub/commit/c6857ebafd51bf198764aff6df2166a083e83b45))
 
 ## [2.23.7](https://github.com/dhis2/app-hub/compare/v2.23.6...v2.23.7) (2021-07-05)